+++ doc/changelog.md
@@ -38,6 +38,7 @@ Syntax: `- short text describing the change _(Your Name)_`
 - _()_
 - _()_
 - Enable seeds to be edited and deleted _(Giancarlo & Moritz)_
+- Update UC for timeline and plants layer _(Paul)_
 - _()_
 - _()_
 - _()_